refactor : NotificationBadge 컴포넌트 수정 및 constants 분리
